+++ base/summarysize.jl
@@ -8,7 +8,7 @@ struct SummarySize
     chargeall::Any
 end
 
-_nfields(x::ANY) = length(typeof(x).types)
+_nfields(@nospecialize x) = length(typeof(x).types)
 
 """
     Base.summarysize(obj; exclude=Union{...}, chargeall=Union{...}) -> Int
@@ -20,9 +20,10 @@ Compute the amount of memory used by all unique objects reachable from the argum
 - `chargeall`: specifies the types of objects to always charge the size of all of their
   fields, even if those fields would normally be excluded.
 """
-function summarysize(obj::ANY;
-                     exclude::ANY = Union{DataType, TypeName, Method},
-                     chargeall::ANY = Union{TypeMapEntry, Core.MethodInstance})
+function summarysize(obj;
+                     exclude = Union{DataType, TypeName, Method},
+                     chargeall = Union{TypeMapEntry, Core.MethodInstance})
+    @nospecialize obj exclude chargeall
     ss = SummarySize(ObjectIdDict(), Any[], Int[], exclude, chargeall)
     size::Int = ss(obj)
     while !isempty(ss.frontier_x)
@@ -61,9 +62,9 @@ function summarysize(obj::ANY;
     return size
 end
 
-(ss::SummarySize)(obj::ANY) = _summarysize(ss, obj)
+(ss::SummarySize)(@nospecialize obj) = _summarysize(ss, obj)
 # define the general case separately to make sure it is not specialized for every type
-@noinline function _summarysize(ss::SummarySize, obj::ANY)
+@noinline function _summarysize(ss::SummarySize, @nospecialize obj)
     key = pointer_from_objref(obj)
     haskey(ss.seen, key) ? (return 0) : (ss.seen[key] = true)
     if _nfields(obj) > 0