+++ test/inference.jl
@@ -270,7 +270,7 @@ for code in Any[
         @code_typed(f18679())[1]
         @code_typed(g18679())[1]]
     @test all(x->isa(x, Type), code.slottypes)
-    local notconst(other::ANY) = true
+    local notconst(@nospecialize(other)) = true
     notconst(slot::TypedSlot) = @test isa(slot.typ, Type)
     function notconst(expr::Expr)
         @test isa(expr.typ, Type)
@@ -443,7 +443,7 @@ function is_typed_expr(e::Expr)
     end
     return false
 end
-test_inferred_static(other::ANY) = true
+test_inferred_static(@nospecialize(other)) = true
 test_inferred_static(slot::TypedSlot) = @test isleaftype(slot.typ)
 function test_inferred_static(expr::Expr)
     if is_typed_expr(expr)
@@ -664,7 +664,7 @@ end
 
 # issue #20704
 f20704(::Int) = 1
-Base.@pure b20704(x::ANY) = f20704(x)
+Base.@pure b20704(@nospecialize(x)) = f20704(x)
 @test b20704(42) === 1
 @test_throws MethodError b20704(42.0)
 
@@ -676,7 +676,7 @@ v20704() = Val{b20704(Any[1.0][1])}
 @test Base.return_types(v20704, ()) == Any[Type{Val{1}}]
 
 Base.@pure g20704(::Int) = 1
-h20704(x::ANY) = g20704(x)
+h20704(@nospecialize(x)) = g20704(x)
 @test g20704(1) === 1
 @test_throws MethodError h20704(1.2)
 
@@ -859,9 +859,9 @@ let f, m
 end
 
 # issue #22290
-f22290() = return nothing
+f22290() = return 3
 for i in 1:3
-    ir = sprint(io->code_llvm(io, f22290, Tuple{}))
+    ir = sprint(io -> code_llvm(io, f22290, Tuple{}))
     @test contains(ir, "julia_f22290")
 end
 
@@ -957,7 +957,7 @@ g22364(x) = f22364(x, Any[[]][1]...)
 @test @inferred(g22364(1)) === 0
 @test @inferred(g22364("1")) === 0.0
 
-function get_linfo(f::ANY, t::ANY)
+function get_linfo(@nospecialize(f), @nospecialize(t))
     if isa(f, Core.Builtin)
         throw(ArgumentError("argument is not a generic function"))
     end
@@ -974,7 +974,7 @@ function get_linfo(f::ANY, t::ANY)
                  (Any, Any, Any, UInt), meth, tt, env, world)
 end
 
-function test_const_return(f::ANY, t::ANY, val::ANY)
+function test_const_return(@nospecialize(f), @nospecialize(t), @nospecialize(val))
     linfo = get_linfo(f, t)
     # If coverage is not enabled, make the check strict by requiring constant ABI
     # Otherwise, check the typed AST to make sure we return a constant.
@@ -1011,6 +1011,29 @@ function test_const_return(f::ANY, t::ANY, val::ANY)
     end
 end
 
+function find_call(code, func, narg)
+    for ex in code
+        isa(ex, Expr) || continue
+        ex = ex::Expr
+        if ex.head === :call && length(ex.args) == narg
+            farg = ex.args[1]
+            if isa(farg, GlobalRef)
+                farg = farg::GlobalRef
+                if isdefined(farg.mod, farg.name) && isconst(farg.mod, farg.name)
+                    farg = getfield(farg.mod, farg.name)
+                end
+            end
+            if farg === func
+                return true
+            end
+        elseif Core.Inference.is_meta_expr(ex)
+            continue
+        end
+        find_call(ex.args, func, narg) && return true
+    end
+    return false
+end
+
 test_const_return(()->1, Tuple{}, 1)
 test_const_return(()->sizeof(Int), Tuple{}, sizeof(Int))
 test_const_return(()->sizeof(1), Tuple{}, sizeof(Int))
@@ -1019,37 +1042,26 @@ test_const_return(()->sizeof(1 < 2), Tuple{}, 1)
 @eval test_const_return(()->Core.sizeof($(Array{Int}())), Tuple{}, sizeof(Int))
 @eval test_const_return(()->Core.sizeof($(Matrix{Float32}(2, 2))), Tuple{}, 4 * 2 * 2)
 
-function find_core_sizeof_call(code)
-    for ex in code
-        isa(ex, Expr) || continue
-        ex = ex::Expr
-        if ex.head === :call && length(ex.args) == 2
-            if ex.args[1] === Core.sizeof || ex.args[1] == GlobalRef(Core, :sizeof)
-                return true
-            end
-        elseif Core.Inference.is_meta_expr(ex)
-            continue
-        end
-        find_core_sizeof_call(ex.args) && return true
-    end
-    return false
-end
-
 # Make sure Core.sizeof with a ::DataType as inferred input type is inferred but not constant.
 function sizeof_typeref(typeref)
     Core.sizeof(typeref[])
 end
 @test @inferred(sizeof_typeref(Ref{DataType}(Int))) == sizeof(Int)
-@test find_core_sizeof_call(first(@code_typed sizeof_typeref(Ref{DataType}())).code)
+@test find_call(first(@code_typed sizeof_typeref(Ref{DataType}())).code, Core.sizeof, 2)
 # Constant `Vector` can be resized and shouldn't be optimized to a constant.
 const constvec = [1, 2, 3]
 @eval function sizeof_constvec()
     Core.sizeof($constvec)
 end
 @test @inferred(sizeof_constvec()) == sizeof(Int) * 3
-@test find_core_sizeof_call(first(@code_typed sizeof_constvec()).code)
+@test find_call(first(@code_typed sizeof_constvec()).code, Core.sizeof, 2)
 push!(constvec, 10)
 @test @inferred(sizeof_constvec()) == sizeof(Int) * 4
+
+test_const_return((x)->isdefined(x, :re), Tuple{Complex128}, true)
+isdefined_f3(x) = isdefined(x, 3)
+@test @inferred(isdefined_f3(())) == false
+@test find_call(first(code_typed(isdefined_f3, Tuple{Tuple{Vararg{Int}}})[1]).code, isdefined, 3)
 
 let isa_tfunc = Core.Inference.t_ffunc_val[
         findfirst(Core.Inference.t_ffunc_key, isa)][3]