+++ test/meta.jl
@@ -144,3 +144,21 @@ baremodule B
 end
 @test B.x == 3
 @test B.M.x == 4
+
+# specialization annotations
+
+function _nospec_some_args(@nospecialize(x), y, @nospecialize z::Int)
+end
+@test first(methods(_nospec_some_args)).nospecialize == 5
+@test first(methods(_nospec_some_args)).sig == Tuple{typeof(_nospec_some_args),Any,Any,Int}
+function _nospec_some_args2(x, y, z)
+    @nospecialize x y
+    return 0
+end
+@test first(methods(_nospec_some_args2)).nospecialize == 3
+function _nospec_with_default(@nospecialize x = 1)
+    2x
+end
+@test collect(methods(_nospec_with_default))[2].nospecialize == 1
+@test _nospec_with_default() == 2
+@test _nospec_with_default(10) == 20